requirements.txt
flask
flask-cors
pyyaml
requests
gunicorn
